network_security_ai/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
jinja2==3.1.2
google-generativeai==0.3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
plotly==5.14.1
python-dotenv==1.0.0
aiohttp==3.8.4
